Move obstacle files to a separate folder.
