+++ source/soccer/obstacle/obstacle.cpp
