+++ source/soccer/obstacle/obstacle.h
