+++ source/soccer/obstacle/obstacle_circle.cpp
