+++ source/soccer/obstacle/obstacle_map.cpp
