+++ source/soccer/obstacle/obstacle_new.h
