+++ source/soccer/obstacle/obstacle_rect.cpp
